src/Twitter_ui/requirements.txt
email-validator==1.1.3
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.0
WTForms==3.0.0
wtforms-html5==0.2.2
config==0.5.1